[fact] Put content merging into separate class #43
